move styles around
