--- frontend/src/charts/ChoroplethMap.module.scss
@@ -1,6 +1,0 @@
-@import "../styles/variables.module";
-
-.ScreenReaderOnly {
-  position: absolute;
-  left: -9999px;
-}
