+++ frontend/src/charts/TableChart.tsx
@@ -28,7 +28,7 @@ import { Tooltip } from "@material-ui/core";
 import WarningRoundedIcon from "@material-ui/icons/WarningRounded";
 import TableContainer from "@material-ui/core/TableContainer";
 import Table from "@material-ui/core/Table";
-import styles from "./TableChart.module.scss";
+import styles from "./Chart.module.scss";
 
 export const MAX_NUM_ROWS_WITHOUT_PAGINATION = 20;
